Finished Restaurant

+++ Java-OOP/Inheritance/Restaurant/src/Dessert.java
@@ -1,8 +1,10 @@
+import java.math.BigDecimal;
+
 public class Dessert extends Food {
 
     private double calories;
 
-    public Dessert(String name, double price, double grams, double calories) {
+    public Dessert(String name, BigDecimal price, double grams, double calories) {
         super(name, price, grams);
         this.calories = calories;
     }
