+++ Java-OOP/Inheritance/Restaurant/src/MainDish.java
@@ -1,6 +1,8 @@
+import java.math.BigDecimal;
+
 public class MainDish extends Food {
 
-    public MainDish(String name, double price, double grams) {
+    public MainDish(String name, BigDecimal price, double grams) {
         super(name, price, grams);
     }
 }
