+++ Java-OOP/Inheritance/Restaurant/src/Starter.java
@@ -1,6 +1,8 @@
+import java.math.BigDecimal;
+
 public class Starter extends Food{
 
-    public Starter(String name, double price, double grams) {
+    public Starter(String name, BigDecimal price, double grams) {
         super(name, price, grams);
     }
 }
